update species key for a few manual corrections
and helper functions; related to #1 @bselden and @jwmorley you might want to be aware of how I did this, and how it differs slightly from the video I link in Issue #1. Basically what I changed is something I already pointed out in the video: I wrote a function to avoid introducing inconsistencies.